requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.14
langchain-openai==0.1.1
python-dotenv==1.0.1
yfinance==0.2.37
pydantic==2.6.4